kmem/slab: check for objects that are free and in array cache
Objects residing in array caches are allocated from the point of slab. If they
are marked as free, it suggests that a double free has happened. Check for
this. This also actually triggers the kmem array cache population from the
checking path, so it can also detect array cache duplicities.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>